Re-organize the files and classes
